fix: hotfix for tmf url (#86)
The `externalSettlementsEndpoint` is very specific to the TMF implementation. This fix will be merged into master to deal with the technical-debt, and the next major version should include a re-factor to fix this in a general manner. I have updated the following story https://github.com/mojaloop/project/issues/2639 to address this going forward.

Fixes:
- bump to patch version
- added missing path for TMF service
- added .nvmrc